+++ lib/karafka/capistrano.rb
@@ -1,2 +1,0 @@
-# Load Capistrano tasks only if Capistrano is loaded
-load File.expand_path('../capistrano/karafka.cap', __FILE__) if defined?(Capistrano)