+++ spec/lib/karafka/capistrano_spec.rb
@@ -1,2 +1,0 @@
-# This file only loads proper rake tasks in Capistrano context
-# No specs here