+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.5.0.2-head
 - #132 - When Kafka is gone, should reconnect after a time period
 - #136 - new ruby-kafka version + other gem dumps
+- ruby-kafka update
 
 ## 0.5.0.2
 - Gems update x3