+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - #136 - new ruby-kafka version + other gem dumps
 - ruby-kafka update
 - #135 - NonMatchingRouteError - better error description in the code
+- #140 - Move Capistrano Karafka to a different specific gem
 
 ## 0.5.0.2
 - Gems update x3