+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - #136 - new ruby-kafka version + other gem dumps
 - ruby-kafka update
 - #135 - NonMatchingRouteError - better error description in the code
+- #140 - Move Capistrano Karafka to a different specific gem
 - #110 - Add call method on a responder class to alias instance build and call
 - #76 - Configs validator
 - #138 - Possibility to have no worker class defined if inline_mode is being used